+++ CHANGELOG.md
@@ -2,3 +2,5 @@
 * Make Changelog
 * Fix documentation comments
 * Write tests, debug collect\_as\_vector
+* Add macro the the modify branch
+* Add encompassing